bot.py
import os
import asyncio
from dotenv import load_dotenv
from aiogram import Bot, Dispatcher
from handlers import links, help

load_dotenv()
BOT_TOKEN = os.getenv("BOT_TOKEN")

bot = Bot(token=BOT_TOKEN)
dp = Dispatcher()
dp.include_router(help.router)
dp.include_router(links.router)

async def main():
    try:
        await dp.start_polling(bot)
    finally:
        await bot.session.close()

if __name__ == "__main__":
    asyncio.run(main())
