requirements.txt
aiogram>=3.1.1
python-dotenv>=1.0
yt-dlp>=2025.9.1
aiohttp>=3.8.5
